DB Entity相關新增

+++ PSI.Core/Entities/PurchaseWeightNote.cs
@@ -7,20 +7,20 @@ namespace PSI.Core.Entities
     public class PurchaseWeightNote : Entity
     {
         public string CarNo { get; set; }
-        public float FullWeight { get; set; }
+        public double FullWeight { get; set; }
         public DateTime FullWeightTime { get; set; }
         public string ScaleNo { get; set; }
         public string InputType { get; set; }
         public string FacNo { get; set; }
         public string Ingredient { get; set; }
-        public float DefectiveWeight { get; set; }
+        public double DefectiveWeight { get; set; }
         public string DefectiveReason { get; set; }
         public DateTime ExcavatorOpTime { get; set; }
         public string ExcavatorOpEmpNo { get; set; }
-        public float CarWeight { get; set; }
+        public double CarWeight { get; set; }
         public DateTime CarWeightTime { get; set; }
-        public float TradeWeight { get; set; }
-        public float FinalDefectiveWeight { get; set; }
+        public double TradeWeight { get; set; }
+        public double FinalDefectiveWeight { get; set; }
         public decimal UnitPrice { get; set; }
         public decimal WantPrice { get; set; }
         public decimal ActualPrice { get; set; }
@@ -30,7 +30,7 @@ namespace PSI.Core.Entities
 
         public decimal TraficFee{ get; set; }   // 派車去指定地方載的運費
         public string PayType { get; set; }
-        public int CustomerId { get; set; }
+        public long CustomerId { get; set; }
         public string CustomerName { get; set; }  // 當下客戶名
         public string NoteStatus { get; set; }
         public string Remark { get; set; }
